# tests/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaskDesk API",
    "version": "1.0.0",
    "description": "TaskDesk Project Management API - Manage projects, tasks, labels, and more"
  },
  "servers": [
    {
      "url": "https://cloud.taskdesk.app/api",
      "description": "TaskDesk API Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key or session token (Bearer)"
      }
    },
    "schemas": {
...
      "ProjectListItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Project"
          },
          {
            "type": "object",
            "properties": {
              "statistics": {
                "$ref": "#/components/schemas/ProjectStatistics"
              },
              "archivedTasks": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BoardTask"
                },
                "description": "Always empty."
              },
              "plannedTasks": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BoardTask"
                },
                "description": "Always empty."
              },
              "columns": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BoardColumn"
                },
                "description": "Always empty."
              }
            },
            "required": [
              "statistics",
              "archivedTasks",
              "plannedTasks",
              "columns"
            ]
          }
        ]
      },
      "ProjectStatistics": {
        "type": "object",
        "properties": {
          "completionPercentage": {
            "type": "number"
          },
          "totalTasks": {
            "type": "number"
          },
          "dueDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "The soonest due date among the project's open tasks."
          }
        },
        "required": [
          "completionPercentage",
          "totalTasks",
          "dueDate"
        ]
      },
      "BoardTask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "number": {
            "type": [
              "number",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "description": "One of: no-priority, low, medium, high, urgent."
          },
          "startDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "dueDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "position": {
            "type": [
              "number",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": [
              "string",
              "null"
            ]
          },
          "assigneeName": {
            "type": [
              "string",
              "null"
            ]
          },
          "assigneeId": {
            "type": [
              "string",
              "null"
            ]
          },
          "assigneeImage": {
            "type": [
              "string",
              "null"
            ]
          },
          "projectId": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskLabel"
            }
          },
          "externalLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskExternalLink"
            }
          }
        },
        "required": [
          "id",
          "title",
          "number",
          "description",
          "status",
          "priority",
          "startDate",
          "dueDate",
          "position",
          "createdAt",
          "userId",
          "assigneeName",
          "assigneeId",
          "assigneeImage",
          "projectId",
          "labels",
          "externalLinks"
        ]
      },
      "TaskLabel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ]
      },
      "TaskExternalLink": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "resourceType": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {},
            "description": "Provider-specific payload, already parsed from the stored JSON string."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "taskId",
          "resourceType",
          "externalId",
          "url",
          "title",
          "metadata",
          "createdAt",
          "updatedAt"
        ]
      },
      "BoardColumn": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The column slug, same as `slug`."
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "isFinal": {
            "type": "boolean"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardTask"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "icon",
          "isFinal",
          "tasks"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "Short prefix used in task identifiers, e.g. KAN-12."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "archivedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Non-null once archived; archived projects are hidden by default."
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Non-null once soft-deleted; deleted projects are excluded everywhere in ordinary use, independent of archivedAt."
          },
          "purgeAfter": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Non-null once soft-deleted: the point after which the project becomes eligible for purge. No purge job exists yet (#198)."
          },
          "position": {
            "type": "number",
            "description": "Sidebar order, ascending."
          },
          "lastTaskNumber": {
            "type": "number",
            "description": "Highest task number issued in this project; the next task gets this plus one."
          }
        },
        "required": [
          "id",
          "workspaceId",
          "slug",
          "icon",
          "name",
          "description",
          "createdAt",
          "archivedAt",
          "deletedAt",
          "purgeAfter",
          "position",
          "lastTaskNumber"
        ]
      },
...
      "WorkItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          },
          "typeId": {
            "type": "string"
          },
          "number": {
            "type": "number",
            "description": "Per-project sequence number. The key is `{project.slug}-{number}`."
          },
          "key": {
            "type": "string",
            "description": "e.g. PROJ-123. Permanent."
          },
          "title": {
            "type": "string"
          },
          "description": {},
          "stateId": {
            "type": "string"
          },
          "priority": {
            "type": [
              "string",
              "null"
            ],
            "description": "One of: low, medium, high, urgent."
          },
          "assigneeId": {
            "type": [
              "string",
              "null"
            ]
          },
          "requesterId": {
            "type": [
              "string",
              "null"
            ]
          },
          "parentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": "string",
            "description": "numeric(20,10) fractional rank, as a string."
          },
          "customerVisibility": {
            "type": "string",
            "description": "One of: private, organisation."
          },
          "archivedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "projectId",
          "workspaceId",
          "typeId",
          "number",
          "key",
          "title",
          "stateId",
          "priority",
          "assigneeId",
          "requesterId",
          "parentId",
          "position",
          "customerVisibility",
          "archivedAt",
          "deletedAt",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/project": {
      "get": {
        "operationId": "listProjects",
        "tags": [
          "Projects"
        ],
        "summary": "List projects",
        "description": "List a workspace's projects in sidebar order, each with rollup task statistics. Archived projects are excluded unless includeArchived is set.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workspaceId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Pass \"true\" to include archived projects in the list."
            },
            "required": false,
            "description": "Pass \"true\" to include archived projects in the list.",
            "name": "includeArchived",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectListItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Workspace ID could not be determined",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "403": {
            "description": "No access to the workspace",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "tags": [
          "Projects"
        ],
        "summary": "Create project",
        "description": "Create a project in a workspace. The slug becomes the prefix of its task identifiers.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "workspaceId": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "workspaceId",
                  "icon",
                  "slug"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, or workspace ID could not be determined",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "403": {
            "description": "No workspace access, or missing project:create permission",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/work-items": {
      "post": {
        "operationId": "createWorkItem",
        "tags": [
          "Work items"
        ],
        "summary": "Create work item",
        "description": "Create a work item in a project. The key is `{project.slug}-{number}`, assigned atomically. The initial state is the project's own default state.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "typeId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "description": {},
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high",
                      "urgent"
                    ]
                  }
                },
                "required": [
                  "typeId",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created work item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, unknown/cross-workspace type, or the project has no default state",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "403": {
            "description": "No workspace access, or missing work_item:create permission",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listWorkItems",
        "tags": [
          "Work items"
        ],
        "summary": "List work items",
        "description": "List a project's work items, oldest first by number. Archived and deleted items are excluded. Not paginated in this first slice.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The project's work items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "403": {
            "description": "No workspace access, or missing work_item:read permission",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/work-items/{key}": {
      "get": {
        "operationId": "getWorkItem",
        "tags": [
          "Work items"
        ],
        "summary": "Get work item",
        "description": "Get a single work item by its permanent key, e.g. PROJ-123.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "key",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The work item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "403": {
            "description": "No workspace access, or missing work_item:read permission",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Work item not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}